[Merge] integrate changes from 2.4 branch (r4381)

+++ branches/avidemux_2.5_branch_gruntster/avidemux/ADM_outputs/oplug_avi/op_savesmart.hxx
@@ -30,9 +30,9 @@
        				uint32_t 	encoderReady;
               			uint32_t 	compEngaged;
 	virtual 		uint8_t 	setupVideo( char *name  );
- 	virtual 		uint8_t 	writeVideoChunk(uint32_t frame );
-				uint8_t		 writeVideoChunk_recode (uint32_t frame);
-				uint8_t		 writeVideoChunk_copy (uint32_t frame,uint32_t first=0);
+ 	virtual 		int 	writeVideoChunk(uint32_t frame );
+				int		 writeVideoChunk_recode (uint32_t frame);
+				int		 writeVideoChunk_copy (uint32_t frame,uint32_t first=0);
               			encoder 	*_encoder;
               			uint8_t 	_cqReenc          ;
